Code/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
opencv-python>=4.8.0
openslide-python>=1.3.1
scikit-learn>=1.3.0
tqdm>=4.65.0
numpy>=1.24.0
